Refactor AuthenticationHelper
req_50
