+++ sernet.gs.service/src/sernet/verinice/interfaces/IAuthService.java
@@ -79,4 +79,5 @@ public interface IAuthService {
 	
 	public boolean isDeactivated();
 
+    public boolean currentUserHasRole(String[] allowedRoles);
 }
