+++ sernet.gs.ui.rcp.main/src/sernet/gs/ui/rcp/main/service/crudcommands/LoadPermissions.java
@@ -24,8 +24,9 @@ import java.util.Set;
 import org.apache.log4j.Logger;
 
 import sernet.gs.common.ApplicationRoles;
-import sernet.gs.ui.rcp.main.service.AuthenticationHelper;
+import sernet.gs.ui.rcp.main.service.ServiceFactory;
 import sernet.verinice.interfaces.GenericCommand;
+import sernet.verinice.interfaces.IAuthService;
 import sernet.verinice.interfaces.IBaseDao;
 import sernet.verinice.model.common.CnATreeElement;
 import sernet.verinice.model.common.Permission;
@@ -55,7 +56,7 @@ public class LoadPermissions extends GenericCommand {
         cte = dao.findById(cte.getDbId());
         permissions = cte.getPermissions();
 
-        boolean isLocalAdmin = AuthenticationHelper.getInstance().currentUserHasRole(new String[] { ApplicationRoles.ROLE_LOCAL_ADMIN });
+        boolean isLocalAdmin = getAuthService().currentUserHasRole(new String[] { ApplicationRoles.ROLE_LOCAL_ADMIN });
         Set<Permission> filteredPermissions = new HashSet<Permission>(permissions.size());
 
         // Hydrate and filter the permissions
@@ -77,5 +78,9 @@ public class LoadPermissions extends GenericCommand {
 
     public Set<Permission> getPermissions() {
         return permissions;
+    }
+
+    private IAuthService getAuthService() {
+        return ServiceFactory.lookupAuthService();
     }
 }
