+++ sernet.gs.ui.rcp.main/src/sernet/verinice/rcp/account/GroupPage.java
@@ -24,10 +24,10 @@ import java.util.Set;
 import org.apache.log4j.Logger;
 import org.eclipse.swt.widgets.Composite;
 
-import sernet.gs.ui.rcp.main.service.AuthenticationHelper;
 import sernet.gs.ui.rcp.main.service.ServiceFactory;
 import sernet.verinice.interfaces.ApplicationRoles;
 import sernet.verinice.interfaces.IAccountService;
+import sernet.verinice.interfaces.IAuthService;
 import sernet.verinice.model.common.accountgroup.AccountGroup;
 import sernet.verinice.model.common.configuration.Configuration;
 
@@ -62,7 +62,7 @@ public class GroupPage extends BaseWizardPage {
         setMessage(Messages.GroupPage_2);
         
         groupWidget = new AccountGroupMultiselectWidget(composite, account);
-        final boolean isLocalAdmin = AuthenticationHelper.getInstance().currentUserHasRole(new String[] { ApplicationRoles.ROLE_LOCAL_ADMIN });
+        final boolean isLocalAdmin = getAuthService().currentUserHasRole(new String[] { ApplicationRoles.ROLE_LOCAL_ADMIN });
         groupWidget.setEnabled(!isLocalAdmin);
     }
 
@@ -134,6 +134,7 @@ public class GroupPage extends BaseWizardPage {
         return ServiceFactory.lookupAccountService();
     }
 
-    
-
+    private IAuthService getAuthService() {
+        return ServiceFactory.lookupAuthService();
+    }
 }
